# docs/api/api.yaml
...
openapi: 3.0.3
info:
  title: Shelfinity Library Management System API
  description: >
    RESTful API for managing library books, users, borrow/return requests,
    reservations, overdue tracking, email configuration, and reports.
    Authentication is delegated entirely to Keycloak (OIDC) — this API validates
    bearer tokens, it does not issue them or store credentials.
  version: 1.0.0
  contact:
    name: Shadow-Codex
servers:
  - url: http://localhost:9080/api
    description: Local development server
...
paths:
...
  /books:
    get:
      tags: [Books]
      summary: List books
      parameters:
        - name: availableOnly
          in: query
          schema: { type: boolean }
        - name: genre
          in: query
          description: "Filter by genre (case-insensitive exact match) — SPEC.md §10.4, resolved"
          schema: { type: string }
      responses:
        '200':
          description: Books retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BookResponse' }
    post:
      tags: [Books]
      summary: Create a book (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookRequest' }
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A book with this ISBN already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /books/search:
    get:
      tags: [Books]
      summary: Search books by title or author
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BookResponse' }
        '400':
          description: Search query is required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /books/bulk-upload:
    post:
      tags: [Books]
      summary: Bulk-upload books from a CSV file (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Upload processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkUploadResult' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /books/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidId'
    get:
      tags: [Books]
      summary: Get a book by ID
      responses:
        '200':
          description: Book retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Books]
      summary: Update a book (admin only)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookRequest' }
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Books]
      summary: Delete a book (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Book deleted }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- Queue
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued OIDC access token.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Forbidden:
      description: Authenticated but not permitted (e.g. admin-only endpoint)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
...
    BookResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        isbn: { type: string }
        description: { type: string }
        genre:
          type: string
          description: "SPEC.md §10.4, resolved"
        publicationYear:
          type: integer
          description: "SPEC.md §10.4, resolved"
        totalCopies: { type: integer }
        availableCopies: { type: integer }
        available: { type: boolean }
        createdAt: { type: string, format: date-time }
    CreateBookRequest:
      type: object
      required: [title, author, totalCopies]
      properties:
        title: { type: string }
        author: { type: string }
        isbn: { type: string }
        description: { type: string }
        genre:
          type: string
          description: "SPEC.md §10.4, resolved"
        publicationYear:
          type: integer
          description: "SPEC.md §10.4, resolved"
        totalCopies: { type: integer, minimum: 1 }
    BulkUploadResult:
      type: object
      properties:
        totalRows: { type: integer }
        successCount: { type: integer }
        errorCount: { type: integer }
        errors:
          type: array
          items: { type: string }
...